test(rest): move the oauth2 session test next to its module
